feat: unify apps/web with docs site, example, and landing page
- Rename apps/docs -> apps/web (@sandagent/web)
- Move sandagent-example content into /example route
- Add fumadocs documentation site at /docs
- Add SandAgent landing page at /
- Delete apps/sandagent-example
- Update deploy workflow for apps/web
- Rewrite @sandagent/sdk README for npm
