--- apps/sandagent-example/app/globals.css
@@ -1,61 +1,0 @@
-@import "tailwindcss";
-@import "kui/styles.css";
-@source "../../../packages/kui";
-
-/* Full height layout */
-html, body {
-  height: 100%;
-}
-
-/* Bounce animation for tool approval UI */
-@keyframes bounce-attention {
-  0%, 100% { 
-    transform: translateY(0) scale(1);
-  }
-  25% { 
-    transform: translateY(-8px) scale(1.02);
-  }
-  50% { 
-    transform: translateY(0) scale(1);
-  }
-}
-
-@keyframes border-pulse {
-  0%, 100% {
-    border-color: hsl(var(--primary));
-    box-shadow: 0 0 0 0 hsla(var(--primary), 0.4);
-  }
-  50% {
-    border-color: hsl(var(--primary));
-    box-shadow: 0 0 0 4px hsla(var(--primary), 0);
-  }
-}
-
-.bounce-animation {
-  animation: bounce-attention 0.6s ease-in-out, border-pulse 2s ease-in-out infinite;
-  border-color: hsl(var(--primary)) !important;
-}
-
-/* Ensure scrollbars are visible */
-.scrollable-content {
-  scrollbar-width: thin;
-  scrollbar-color: hsl(var(--muted-foreground)) hsl(var(--muted));
-}
-
-.scrollable-content::-webkit-scrollbar {
-  width: 8px;
-  height: 8px;
-}
-
-.scrollable-content::-webkit-scrollbar-track {
-  background: hsl(var(--muted));
-}
-
-.scrollable-content::-webkit-scrollbar-thumb {
-  background-color: hsl(var(--muted-foreground));
-  border-radius: 4px;
-}
-
-.scrollable-content::-webkit-scrollbar-thumb:hover {
-  background-color: hsl(var(--foreground) / 0.5);
-}
