--- apps/sandagent-example/app/layout.tsx
@@ -1,24 +1,0 @@
-import type { Metadata } from "next";
-import { ThemeProvider } from "next-themes";
-import "./globals.css";
-
-export const metadata: Metadata = {
-  title: "SandAgent Chat",
-  description: "AI Chat example using SandAgent with KUI",
-};
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang="en" suppressHydrationWarning>
-      <body>
-        <ThemeProvider attribute="class" defaultTheme="dark" enableSystem>
-          {children}
-        </ThemeProvider>
-      </body>
-    </html>
-  );
-}
