--- apps/sandagent-example/next.config.js
@@ -1,13 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  transpilePackages: [
-    "@sandagent/core",
-    "@sandagent/sdk",
-    "@sandagent/sandbox-e2b",
-    "@sandagent/sandbox-sandock",
-    "@sandagent/sandbox-daytona",
-    "kui",
-  ],
-};
-
-module.exports = nextConfig;
