--- apps/sandagent-example/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@sandagent/example",
-  "version": "0.1.0",
-  "private": true,
-  "description": "Complete SandAgent example with AI Chat UI, template selector, and real-time streaming",
-  "scripts": {
-    "dev": "concurrently \"pnpm dev:deps\" \"next dev\"",
-    "dev:deps": "pnpm --filter @sandagent/manager --filter @sandagent/sandbox-e2b --filter @sandagent/sandbox-sandock --filter @sandagent/sandbox-daytona --parallel dev",
-    "dev:next": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@ai-sdk/react": "^3.0.3",
-    "@sandagent/sdk": "workspace:*",
-    "@sandagent/sandbox-daytona": "workspace:*",
-    "@sandagent/sandbox-e2b": "workspace:*",
-    "@sandagent/sandbox-sandock": "workspace:*",
-    "ai": "^6.0.19",
-    "kui": "workspace:*",
-    "lucide-react": "^0.556.0",
-    "next": "16.0.10",
-    "next-themes": "^0.4.6",
-    "react": "^19.2.0",
-    "react-dom": "^19.2.0"
-  },
-  "devDependencies": {
-    "@tailwindcss/postcss": "^4.1.17",
-    "@types/node": "^20.10.0",
-    "@types/react": "^19.2.0",
-    "@types/react-dom": "^19.2.0",
-    "concurrently": "^8.2.2",
-    "tailwindcss": "^4.1.17",
-    "typescript": "^5.9.3"
-  }
-}
