+++ apps/web/app/(example)/layout.tsx
@@ -1,0 +1,15 @@
+"use client";
+
+import { ThemeProvider } from "next-themes";
+
+export default function ExampleLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <ThemeProvider attribute="class" defaultTheme="dark" enableSystem>
+      {children}
+    </ThemeProvider>
+  );
+}
