+++ apps/web/app/api/chat/route.ts
