+++ apps/web/app/layout.tsx
@@ -1,0 +1,27 @@
+import { RootProvider } from 'fumadocs-ui/provider/next';
+import './global.css';
+import { Inter } from 'next/font/google';
+import type { Metadata } from 'next';
+
+const inter = Inter({
+  subsets: ['latin'],
+});
+
+export const metadata: Metadata = {
+  title: {
+    default: 'SandAgent',
+    template: '%s | SandAgent',
+  },
+  description:
+    'Turn powerful Coding Agents into your product\'s superpower. AI SDK compatible.',
+};
+
+export default function Layout({ children }: LayoutProps<'/'>) {
+  return (
+    <html lang="en" className={inter.className} suppressHydrationWarning>
+      <body className="flex flex-col min-h-screen">
+        <RootProvider>{children}</RootProvider>
+      </body>
+    </html>
+  );
+}
