+++ apps/web/content/docs/meta.json
@@ -1,0 +1,10 @@
+{
+  "title": "Documentation",
+  "pages": [
+    "index",
+    "quick-start",
+    "sandboxes",
+    "templates",
+    "api"
+  ]
+}
