+++ apps/web/lib/cn.ts
@@ -1,0 +1,1 @@
+export { twMerge as cn } from 'tailwind-merge';
