+++ apps/web/lib/example/artifact-processor.ts
