+++ apps/web/package.json
@@ -1,0 +1,43 @@
+{
+  "name": "@sandagent/web",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "build": "next build",
+    "dev": "next dev",
+    "start": "next start",
+    "types:check": "fumadocs-mdx && next typegen && tsc --noEmit",
+    "postinstall": "fumadocs-mdx",
+    "lint": "biome check",
+    "format": "biome format --write"
+  },
+  "dependencies": {
+    "fumadocs-core": "16.5.4",
+    "fumadocs-mdx": "14.2.7",
+    "fumadocs-ui": "16.5.4",
+    "lucide-react": "^0.563.0",
+    "next": "16.1.6",
+    "react": "^19.2.4",
+    "react-dom": "^19.2.4",
+    "tailwind-merge": "^3.4.0",
+    "@ai-sdk/react": "^3.0.3",
+    "@sandagent/sdk": "workspace:*",
+    "@sandagent/sandbox-daytona": "workspace:*",
+    "@sandagent/sandbox-e2b": "workspace:*",
+    "@sandagent/sandbox-sandock": "workspace:*",
+    "ai": "^6.0.19",
+    "kui": "workspace:*",
+    "next-themes": "^0.4.6"
+  },
+  "devDependencies": {
+    "@tailwindcss/postcss": "^4.1.18",
+    "@types/mdx": "^2.0.13",
+    "@types/node": "^25.2.1",
+    "@types/react": "^19.2.13",
+    "@types/react-dom": "^19.2.3",
+    "postcss": "^8.5.6",
+    "tailwindcss": "^4.1.18",
+    "typescript": "^5.9.3",
+    "@biomejs/biome": "^2.3.14"
+  }
+}
